+++ src/01-What_is_Node/README.md
@@ -1,0 +1,28 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+# What is node?
+
+Node.js is a platform built on Chrome’s JavaScript runtime for easily building fast, scalable network applications.
+Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices. The exam will test your basic understanding of Node’s architecture, its main components (APIs, bindings, V8 and libuv), and how they interact.
+
+* [V8 JavaScript engine](http://strongloop.com/developers/videos/#a-video-intro-to-how-node-works)
+* [libuv: cross platform asynchronous I/O for Node](https://github.com/libuv/libuv)
+* [APIs](http://nodejs.org/api)
+
+There are two additional videos worth watching that introduce the fundamentals of Node:
+* [How Node works](http://strongloop.com/developers/videos/#a-video-intro-to-how-node-works)
+* [Node fundamentals](https://strongloop.com/node-js/videos/#a-video-intro-to-nodejs-fundamentals)
+
+## Content
+
+No exercises by the moment sorry :(
+To contrib use [pull requests](https://help.github.com/articles/using-pull-requests).
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
