+++ src/02-Working_with_npm_and_modules/module_1/README.md
@@ -1,0 +1,12 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+Module 1 to show a message when it's required
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
