+++ src/02-Working_with_npm_and_modules/module_1/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "module_2",
+  "name": "module_1",
   "version": "0.0.0",
-  "description": "ERROR: No README.md file found!",
+  "description": "Say hello World",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": "",
   "author": "",
-  "license": "BSD"
+  "license": "MIT"
 }
