+++ src/02-Working_with_npm_and_modules/module_2/README.md
@@ -1,0 +1,12 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+Module 2 to require the Module 1
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
