+++ src/03-Node_Concepts/streams/writable.js
