+++ src/04-Working_with_Express/README.md
@@ -1,0 +1,20 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+# Name of section
+
+The [Express](http://expressjs.com/) framework is arguably one of the most popular npm modules and what many developers start and build applications with.
+What is Express? Express is a minimal and flexible Node-based web application framework, providing a robust set of features for building single and multi-page, and hybrid web applications. For the purposes of the exam you should know how to build a basic web application using Express.
+
+## Content
+
+* ./express.js
+* ./express-domain.js
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
