+++ src/05-Working_with_the_Process_API/README.md
@@ -1,0 +1,19 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+# Name of section
+
+The [process](https://nodejs.org/docs/latest-v0.12.x/api/process.html) object is a global object and can be accessed from anywhere.
+For the purposes of the exam you should understand signal events, accessing environmental variables and dealing with child processes.
+
+## Content
+
+* ./process.js
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
