+++ src/07-Sockets/README.md
@@ -1,0 +1,21 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+# Name of section
+
+[WebSockets](http://strongloop.com/strongblog/real-time-engines-in-node-js/) in Node commonly manifest themselves when developers make use of the [socket.io](http://socket.io) framework.
+[WebSockets](http://strongloop.com/strongblog/real-time-engines-in-node-js/) is a protocol providing full-duplex communications channels over a single TCP connection. 
+For the purposes of the exam, you should have a basic understanding of how [WebSockets](http://strongloop.com/strongblog/real-time-engines-in-node-js/) work, why they are advantageous in Node and how they can be integrated into your application with the [socket.io](http://socket.io) framework.
+
+## Content
+
+* ./websocket.js
+* ./websocket.html
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
