+++ src/08-Clustering_and_Multi-Process/README.md
@@ -1,0 +1,23 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+# Name of section
+
+Before deploying a Node app, it’s important to understand what your options are for scalability not just around performance, but also for uptime. 
+
+For the purposes of the exam you should have a general understanding of how Node’s [Cluster API](https://nodejs.org/docs/latest-v0.12.x/api/cluster.html) works and how it might be leveraged to create scalable applications.
+
+## Content
+
+* ./cluster.js
+  * ./workers/worker1.js
+* ./cluster-domain.js
+  * ./workers/domain-worker.js
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
