+++ src/09-Deployment/README.md
@@ -1,0 +1,21 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+# Deployment
+
+For the purposes of the exam it it’s import to understand some of the fundamental tasks you’ll need to preform in the application deployment process.
+
+For example high-availability, clustering in multi-core environments, zero downtime deployments and log rotation. Qasim Zaidi has put together an excellent [“Node in production”](http://qzaidi.github.io/2013/05/14/node-in-production/) post on these topics.
+
+## Content
+
+No files by the moment sorry :( 
+To contrib use [pull requests](https://help.github.com/articles/using-pull-requests).
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
