+++ src/10-Troubleshooting/README.md
@@ -1,0 +1,22 @@
+node-certification
+==================
+
+SCND Strongloop Certification Exam work out
+
+# Name of section
+
+It is important for Node developers to have an understanding of what development tools are at their disposal and how to leverage them. For the purposes of the exam, you should have a general understanding of the following tools:
+
+* Debugging: Understanding the built in [debugger](https://nodejs.org/docs/latest-v0.12.x/api/debugger.html) and [Node inspector]()
+* Memory leak detection: [Using event listeners](https://nodejs.org/docs/latest-v0.12.x/api/events.html) to catch memory leaks
+* [Using REPL](https://nodejs.org/docs/latest-v0.12.x/api/repl.html): A Read-Eval-Print-Loop (REPL) is available both as a standalone program and easily includable in other programs. The REPL provides a way to interactively run JavaScript and see the results. It can be used for debugging, testing, or just trying things out.
+
+## Content
+
+* ./repl.js
+
+## License
+[Read the LICENSE file (MIT, anyway)](../../LICENSE)
+
+## Reference
+[http://strongloop.com/node-js/certification/scnd-study-guide/](http://strongloop.com/node-js/certification/scnd-study-guide/)
